pyproject.toml
[tool.poetry]
name = "moscowliuda-webinar"
version = "0.1.0"
description = "Automation for @moscowliuda webinar"
authors = ["Anton Demkin <antondemkin@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"
loguru = "^0.6.0"
python-dotenv = "^0.20.0"
yagmail = "^0.15.285"
gspread = "^5.5.0"
pymorphy2 = "^0.9.1"
pyphrasy = "^0.2.0"
pymorphy2-dicts-ru = "^2.4.417127.4579844"
aiohttp = "^3.8.3"
aiohttp-jinja2 = "^1.5"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
mypy = "^0.971"
flake8 = "^5.0.4"
pylint = "^2.15.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.main]
disable = [
    'missing-function-docstring',
    'missing-class-docstring',
    'missing-module-docstring',
    'line-too-long',
    'broad-except',
    'invalid-name',
]
load-plugins = 'pylint.extensions.no_self_use'

[tool.flake8]
max-line-length = 100

[tool.isort]
profile = 'google'

[tool.black]
line-length = 80
